requirements-postgres.txt
# Dependencias para PostgreSQL
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Mail==0.9.1
python-dotenv==1.1.1
reportlab==4.4.3

# Driver PostgreSQL
psycopg2-binary==2.9.9

# Utilidades adicionales
python-dateutil==2.8.2
Werkzeug==3.1.3
SQLAlchemy==2.0.43
